backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pyodbc==5.1.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
azure-storage-blob==12.19.0
azure-storage-queue==12.9.0
aiofiles==23.2.1
httpx==0.27.0
structlog==24.1.0
slowapi==0.1.9
email-validator==2.1.0
orjson==3.9.13
stripe>=14.0.0
